// packages/connection-types-shared/src/concepts/connectionTypes/types.ts
export enum ConnectionTypeFieldType {
  Boolean = 'boolean',
  Dropdown = 'dropdown',
  File = 'file',
  Hidden = 'hidden',
  Numeric = 'numeric',
  Section = 'section',
  ShortText = 'short-text',
  Text = 'text',
  URI = 'uri',
}
// ...
export type ConnectionTypeFieldTypeUnion =
  | ConnectionTypeFieldType
  | 'boolean'
  | 'dropdown'
  | 'file'
  | 'hidden'
  | 'numeric'
  | 'section'
  | 'short-text'
  | 'text'
  | 'uri';

type Field<T extends ConnectionTypeFieldTypeUnion> = {
  type: T;
  name: string;
  description?: string;
};

export type ConnectionTypeCommonProperties<V = string> = {
  defaultValue?: V;
  defaultReadOnly?: boolean;
  deferInput?: boolean;
};

// P default to an empty set of properties
// eslint-disable-next-line @typescript-eslint/ban-types
export type DataField<T extends ConnectionTypeFieldTypeUnion, V = string, P = {}> = Field<T> & {
  envVar: string;
  required?: boolean;
  properties: P & ConnectionTypeCommonProperties<V>;
};

export type SectionField = Field<ConnectionTypeFieldType.Section | 'section'>;

export type HiddenField = DataField<ConnectionTypeFieldType.Hidden | 'hidden'>;